requirements.txt
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.34
PyYAML==6.0
flask
pymysql==1.0.2
pandas